REFACTOR: change kafka service to kafka module in group schedule connector
