+++ kafka/src/main/kotlin/vp/togedo/kafka/service/GroupKafkaService.kt
@@ -2,11 +2,8 @@ package vp.togedo.kafka.service
 
 import reactor.core.publisher.Mono
 import vp.togedo.kafka.data.group.InviteGroupEventDto
-import vp.togedo.kafka.data.groupSchedule.CreateGroupScheduleEventDto
 
 interface GroupKafkaService {
 
     fun publishInviteGroupEvent(inviteGroupEventDto: InviteGroupEventDto): Mono<Void>
-
-    fun publishCreateGroupScheduleEvent(createGroupScheduleEventDto: CreateGroupScheduleEventDto): Mono<Void>
 }
