+++ kafka/src/main/kotlin/vp/togedo/kafka/service/impl/GroupKafkaServiceImpl.kt
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
 import vp.togedo.kafka.data.enums.EventEnums
 import vp.togedo.kafka.data.group.InviteGroupEventDto
-import vp.togedo.kafka.data.groupSchedule.CreateGroupScheduleEventDto
 import vp.togedo.kafka.publish.KafkaPublisher
 import vp.togedo.kafka.service.GroupKafkaService
 
@@ -17,13 +16,6 @@ class GroupKafkaServiceImpl(
         return kafkaPublisher.publishKafkaEvent(
             EventEnums.INVITE_GROUP,
             inviteGroupEventDto
-        )
-    }
-
-    override fun publishCreateGroupScheduleEvent(createGroupScheduleEventDto: CreateGroupScheduleEventDto): Mono<Void> {
-        return kafkaPublisher.publishKafkaEvent(
-            EventEnums.CREATE_GROUP_SCHEDULE,
-            createGroupScheduleEventDto
         )
     }
 }
