+++ kafka/src/main/kotlin/vp/togedo/kafka/service/impl/GroupScheduleKafkaServiceImpl.kt
@@ -3,6 +3,8 @@ package vp.togedo.kafka.service.impl
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
 import vp.togedo.kafka.data.enums.EventEnums
+import vp.togedo.kafka.data.groupSchedule.ConfirmScheduleEventDto
+import vp.togedo.kafka.data.groupSchedule.CreateGroupScheduleEventDto
 import vp.togedo.kafka.data.groupSchedule.SuggestGroupScheduleEventDto
 import vp.togedo.kafka.publish.KafkaPublisher
 import vp.togedo.kafka.service.GroupScheduleKafkaService
@@ -19,10 +21,17 @@ class GroupScheduleKafkaServiceImpl(
         )
     }
 
-    override fun publishConfirmScheduleEvent(confirmScheduleEvent: SuggestGroupScheduleEventDto): Mono<Void> {
+    override fun publishConfirmScheduleEvent(confirmScheduleEvent: ConfirmScheduleEventDto): Mono<Void> {
         return kafkaPublisher.publishKafkaEvent(
             EventEnums.CONFIRM_SCHEDULE,
             confirmScheduleEvent
+        )
+    }
+
+    override fun publishCreateGroupScheduleEvent(createGroupScheduleEventDto: CreateGroupScheduleEventDto): Mono<Void> {
+        return kafkaPublisher.publishKafkaEvent(
+            EventEnums.CREATE_GROUP_SCHEDULE,
+            createGroupScheduleEventDto
         )
     }
 }
